Move to SelectingItemsControl
